backend/requirements-prod.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyMuPDF>=1.24.0
pymongo>=4.6.0
redis>=5.0.0
gunicorn>=21.2.0 